backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["trojrobert <trojrobert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
python-dotenv = "1.0.1"
fastapi = "0.114.2"
uvicorn = "^0.30.6"
openai = "^1.45.0"
langchain = "^0.3.0"
pypdf2 = "^3.0.1"
python-multipart = "^0.0.9"
mangum = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
